move topbar to app component. swap out ion router. add side menu.
