+++ src/app/_components/angular-material.module.ts
@@ -8,7 +8,11 @@ import {
 } from '@angular/material';
 
 import { MatCardModule } from '@angular/material/card';
+import { MatDividerModule } from '@angular/material/divider';
 import { MatExpansionModule } from '@angular/material/expansion';
+import { MatIconModule } from '@angular/material/icon';
+import { MatListModule } from '@angular/material/list';
+import { MatSidenavModule } from '@angular/material/sidenav';
 import { MatSnackBarModule } from '@angular/material/snack-bar';
 import { MatToolbarModule } from '@angular/material/toolbar';
 
@@ -17,7 +21,6 @@ import { ActionSnackBarComponent } from './action-snack-bar/action-snack-bar';
 import { PropertyCardComponent } from './property-card/property-card';
 import { LandlordCardComponent } from './landlord-card/landlord-card';
 import { StarsDisplayComponent } from './stars-display/stars-display';
-import { TopBarComponent } from './top-bar/top-bar';
 
 @NgModule({
   declarations: [
@@ -25,14 +28,17 @@ import { TopBarComponent } from './top-bar/top-bar';
     PropertyCardComponent,
     LandlordCardComponent,
     StarsDisplayComponent,
-    TopBarComponent,
   ],
   imports: [
     MatButtonModule,
     MatCheckboxModule,
     MatInputModule,
     MatCardModule,
+    MatDividerModule,
     MatExpansionModule,
+    MatIconModule,
+    MatListModule,
+    MatSidenavModule,
     MatSnackBarModule,
     MatToolbarModule,
     CommonModule,
@@ -43,14 +49,17 @@ import { TopBarComponent } from './top-bar/top-bar';
     MatCheckboxModule,
     MatInputModule,
     MatCardModule,
+    MatDividerModule,
     MatExpansionModule,
+    MatIconModule,
+    MatListModule,
+    MatSidenavModule,
     MatSnackBarModule,
     MatToolbarModule,
     ActionSnackBarComponent,
     PropertyCardComponent,
     LandlordCardComponent,
     StarsDisplayComponent,
-    TopBarComponent,
   ],
 })
 export class AngularMaterialModule {}
