--- src/app/_components/top-bar/top-bar.scss
@@ -1,16 +1,0 @@
-mat-toolbar.mat-primary {
-  position: relative;
-  z-index: 1;
-}
-
-.account-nav {
-  width: 100%;
-  text-align: right;
-  flex: 1;
-  padding: 0 12px;
-}
-
-button {
-  margin-right: .5rem;
-  margin-left: .5rem;
-}
