+++ src/app/_pages/home/home.page.html
@@ -1,8 +1,4 @@
-<top-bar></top-bar>
-
-<ion-content padding text-center>
-
-  <section>
+  <section class="header">
     <p>Welcome to HousingDB</p>
 
     <h2>Looking for your next place to live?</h2>
@@ -11,10 +7,10 @@
 
     <p>
       <a routerLink="/properties">
-        <button mat-raised-button color="primary">Browse Properties</button>
+        <button mat-raised-button>Browse Properties</button>
       </a>
       <a routerLink="/landlords">
-        <button mat-raised-button color="primary">Browse Landlords</button>
+        <button mat-raised-button>Browse Landlords</button>
       </a>
     </p>
   </section>
@@ -65,5 +61,3 @@
       <button mat-raised-button color="primary" (click)="doLogin()">Sign Up Now</button>
     </a>
   </section>
-
-</ion-content>
