+++ src/app/_pages/home/home.page.scss
@@ -1,3 +1,7 @@
+:host {
+  text-align: center;
+}
+
 .add-types {
   display: flex;
   flex-flow:row wrap;
@@ -9,12 +13,23 @@
 }
 
 section {
-  // max-width: 62.5rem;
   max-width: 82.5rem;
   padding: .5rem;
   margin-left: auto;
   margin-right: auto;
   margin-bottom: 2rem;
+}
+
+section.header {
+  margin-left: -1rem;
+  margin-right: -1rem;
+  width: auto;
+  max-width: unset;
+  padding-top: 3rem;
+  padding-bottom: 3rem;
+  h2 {
+    font-size: 3rem;
+  }
 }
 
 button {
