+++ src/app/_pages/property-add-reviews/property-add-reviews.page.html
@@ -1,12 +1,8 @@
-<top-bar></top-bar>
-
 <mat-toolbar class="page-info">
   <mat-toolbar-row>
     Add Property Review
   </mat-toolbar-row>
 </mat-toolbar>
-
-<ion-content padding>
 
   <form (submit)="submit()" [formGroup]="form" #ngFormDirective="ngForm" [ngClass]="{'submit-attempt': submitAttempt}">
 
@@ -55,5 +51,3 @@
       <button mat-raised-button color="primary" type="submit">Submit</button>
     </div>
   </form>
-
-</ion-content>
