+++ src/app/app.component.scss
@@ -1,0 +1,33 @@
+mat-sidenav-container {
+  height: 100%;
+}
+
+mat-sidenav {
+  width: 18rem;
+  max-width: 100%;
+}
+
+.sidenav-header {
+  height: 64px;
+}
+
+
+.primary-toolbar {
+  position: relative;
+  z-index: 1;
+
+  .account-nav {
+    width: 100%;
+    text-align: right;
+    flex: 1;
+    padding: 0 12px;
+
+    button {
+      margin-right: .5rem;
+      margin-left: .5rem;
+    }
+  }
+}
+
+// .app-content and mat-toolbar moved to global.scss so it could apply to
+// mat-toolbar elements inside child components
